--- language/models/declaration.js
@@ -1,72 +1,0 @@
-
-import Cache from "./cache";
-
-export default class Declaration {
-  /**
-   * 
-   * @param {"procedure"|"subroutine"|"file"|"struct"|"subitem"|"variable"|"constant"|"tag"} type 
-   */
-  constructor(type) {
-    this.type = type;
-    this.name = ``;
-
-    /** @type {import("../parserTypes").Keywords} */
-    this.keyword = {};
-    
-    this.description = ``;
-
-    /** @type {{tag: string, content: string}[]} */
-    this.tags = [];
-
-    /** @type {import("../parserTypes").DefinitionPosition} */
-    this.position = undefined;
-
-    /** @type {import("../parserTypes").Reference[]} */
-    this.references = [];
-
-    // Not used in subitem:
-    /** @type {Declaration[]} */
-    this.subItems = [];
-
-    // Only used in procedure
-    this.readParms = false;
-
-    // Used for subroutines and procedures
-    /** @type {{start: number, end: number}} */
-    this.range = {
-      start: null,
-      end: null
-    };
-
-    // Only used in procedures
-    /** @type {Cache|undefined} */
-    this.scope = undefined;
-  }
-
-  clone() {
-    const clone = new Declaration(this.type);
-    clone.name = this.name;
-    clone.keyword = this.keyword;
-    clone.description = this.description;
-    clone.tags = this.tags;
-
-    if (this.position) {
-      clone.position = {
-        path: this.position.path,
-        line: this.position.line
-      };
-    }
-
-    clone.subItems = this.subItems.map(subItem => subItem.clone());
-    
-    clone.range = {
-      start: this.range.start,
-      end: this.range.end
-    };
-
-    //clone.references = this.references;
-    //clone.readParms = this.readParms;
-    //clone.scope = this.scope;
-    return clone;
-  }
-}
